requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
wordcloud==1.6.0
matplotlib==3.1.3
streamlit==0.84.0
pandas==1.2.3
torch==1.8.1
seaborn==0.10.1
transformers==4.3.3
feel_it==1.0.3
numpy==1.21.0
Pillow==8.3.1
scikit_learn==0.24.2
torch==1.6.0+cpu